requirements.txt
chromadb==0.5.15
google-generativeai==0.8.3
greenlet==3.1.1
grpcio-status==1.67.0
httptools==0.6.4
llama-deploy==0.1.3
openai==1.52.0
opencv-python==4.10.0.84
PyAutoGUI==0.9.53
pygame==2.5.2
reflex-audio-capture==0.0.4
reflex-img-comparison-slider==0.0.1
reflex-webcam==0.0.6
replicate==0.26.0
uvloop==0.21.0
